chore: rename sinks module to destinations
