+++ pg_replicate/src/pipeline/destinations/stdout.rs
